+++ tensorflow_io/core/python/api/experimental/audio.py
@@ -24,4 +24,6 @@ from tensorflow_io.core.python.experimental.audio_ops import (  # pylint: disabl
     encode_vorbis,
     decode_mp3,
     encode_mp3,
+    AudioIOTensor,
+    AudioIODataset,
 )
