+++ tensorflow_io/core/python/experimental/audio_ops.py
@@ -14,6 +14,8 @@
 # ==============================================================================
 """audio"""
 
+import sys
+
 import tensorflow as tf
 
 from tensorflow_io.core.python.ops import core_ops
@@ -167,3 +169,154 @@ def encode_mp3(input, rate, name=None):  # pylint: disable=redefined-builtin
       output: Encoded audio.
     """
     return core_ops.io_audio_encode_mp3(input, rate, name=name)
+
+
+class AudioIOTensor:
+    """AudioIOTensor"""
+
+    # =============================================================================
+    # Constructor
+    # =============================================================================
+    def __init__(self, filename, dtype=None):
+        with tf.name_scope("AudioIOTensor"):
+            if not tf.executing_eagerly():
+                assert dtype is not None, "dtype must be provided in graph mode"
+            if sys.platform == "linux":
+                try:
+                    from tensorflow_io.core.python.ops import (  # pylint: disable=import-outside-toplevel,unused-import
+                        ffmpeg_ops,
+                    )
+                except NotImplementedError:
+                    pass
+            resource = core_ops.io_audio_readable_init(filename)
+            if tf.executing_eagerly():
+                shape, dtype, rate = core_ops.io_audio_readable_spec(resource)
+                dtype = tf.as_dtype(dtype.numpy())
+            else:
+                shape, _, rate = core_ops.io_audio_readable_spec(resource)
+            self._resource = resource
+            self._shape = shape
+            self._dtype = dtype
+            self._rate = rate
+            super().__init__()
+
+    # =============================================================================
+    # Accessors
+    # =============================================================================
+
+    @property
+    def shape(self):
+        """Returns the `TensorShape` that represents the shape of the tensor."""
+        return self._shape
+
+    @property
+    def dtype(self):
+        """Returns the `dtype` of elements in the tensor."""
+        return self._dtype
+
+    @property
+    def rate(self):
+        """The sample `rate` of the audio stream"""
+        return self._rate
+
+    # =============================================================================
+    # String Encoding
+    # =============================================================================
+    def __repr__(self):
+        return "<AudioIOTensor: shape={}, dtype={}, rate={}>".format(
+            self.shape, self.dtype, self.rate
+        )
+
+    # =============================================================================
+    # Tensor Type Conversions
+    # =============================================================================
+
+    def to_tensor(self):
+        """Converts this `IOTensor` into a `tf.Tensor`.
+
+        Args:
+          name: A name prefix for the returned tensors (optional).
+
+        Returns:
+          A `Tensor` with value obtained from this `IOTensor`.
+        1"""
+        return core_ops.io_audio_readable_read(self._resource, 0, -1, dtype=self._dtype)
+
+    # =============================================================================
+    # Indexing and slicing
+    # =============================================================================
+    def __getitem__(self, key):
+        """Returns the specified piece of this IOTensor."""
+        # always convert to tuple to process
+        if not isinstance(key, tuple):
+            key = tuple([key])
+        # get the start and stop of each element
+        indices = [
+            (k.start, k.stop) if isinstance(k, slice) else (k, k + 1) for k in key
+        ]
+        # get the start and stop, and use 0 (start) and -1 (stop) if needed
+        indices = list(zip(*indices))
+        start = [0 if e is None else e for e in indices[0]]
+        stop = [-1 if e is None else e for e in indices[1]]
+
+        item = core_ops.io_audio_readable_read(
+            self._resource, start=start, stop=stop, dtype=self._dtype
+        )
+
+        # in case certain dimension is not slice, then this dimension will need to
+        # collapse as `0`, otherwise `:` or `slice(None, None, None)`
+        indices = [slice(None) if isinstance(k, slice) else 0 for k in key]
+
+        return item.__getitem__(indices)
+
+    def __len__(self):
+        """Returns the total number of items of this IOTensor."""
+        return self._shape[0]
+
+
+class AudioIODataset(tf.data.Dataset):
+    """AudioIODataset"""
+
+    def __init__(self, filename, dtype=None):
+        """AudioIODataset."""
+        with tf.name_scope("AudioIODataset"):
+            if not tf.executing_eagerly():
+                assert dtype is not None, "dtype must be provided in graph mode"
+            if sys.platform == "linux":
+                try:
+                    from tensorflow_io.core.python.ops import (  # pylint: disable=import-outside-toplevel,unused-import
+                        ffmpeg_ops,
+                    )
+                except NotImplementedError:
+                    pass
+            resource = core_ops.io_audio_readable_init(filename)
+            if tf.executing_eagerly():
+                shape, dtype, _ = core_ops.io_audio_readable_spec(resource)
+                dtype = tf.as_dtype(dtype.numpy())
+            else:
+                shape, _, _ = core_ops.io_audio_readable_spec(resource)
+
+            capacity = 1024  # kwargs.get("capacity", 4096)
+
+            self._resource = resource
+            dataset = tf.data.Dataset.range(0, shape[0], capacity)
+            dataset = dataset.map(
+                lambda index: core_ops.io_audio_readable_read(
+                    resource, index, index + capacity, dtype=dtype
+                )
+            )
+            dataset = dataset.apply(
+                tf.data.experimental.take_while(lambda v: tf.greater(tf.shape(v)[0], 0))
+            )
+            dataset = dataset.unbatch()
+            self._dataset = dataset
+            super().__init__(
+                self._dataset._variant_tensor
+            )  # pylint: disable=protected-access
+
+    def _inputs(self):
+        return []
+
+    @property
+    def element_spec(self):
+        return self._dataset.element_spec
