+++ tensorflow_io/core/python/ops/io_tensor.py
@@ -15,9 +15,8 @@
 """IOTensor"""
 
 import tensorflow as tf
-from tensorflow_io.core.python.ops import core_ops
+from tensorflow_io.core.python.experimental import audio_ops
 from tensorflow_io.core.python.ops import io_tensor_ops
-from tensorflow_io.core.python.ops import audio_io_tensor_ops
 from tensorflow_io.core.python.ops import json_io_tensor_ops
 from tensorflow_io.core.python.ops import hdf5_io_tensor_ops
 from tensorflow_io.core.python.ops import kafka_io_tensor_ops
@@ -34,165 +33,165 @@ from tensorflow_io.core.python.ops import arrow_io_tensor_ops
 class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     """IOTensor
 
-  An `IOTensor` is a tensor with data backed by IO operations. For example,
-  an `AudioIOTensor` is a tensor with data from an audio file, a
-  `KafkaIOTensor` is a tensor with data from reading the messages of a Kafka
-  stream server.
+    An `IOTensor` is a tensor with data backed by IO operations. For example,
+    an `AudioIOTensor` is a tensor with data from an audio file, a
+    `KafkaIOTensor` is a tensor with data from reading the messages of a Kafka
+    stream server.
 
-  The `IOTensor` is indexable, supporting `__getitem__()` and
-  `__len__()` methods in Python. In other words, it is a subclass of
-  `collections.abc.Sequence`.
+    The `IOTensor` is indexable, supporting `__getitem__()` and
+    `__len__()` methods in Python. In other words, it is a subclass of
+    `collections.abc.Sequence`.
 
-  Example:
+    Example:
 
-  ```python
-  >>> import tensorflow_io as tfio
-  >>>
-  >>> samples = tfio.IOTensor.from_audio("sample.wav")
-  >>> print(samples[1000:1005])
-  ... tf.Tensor(
-  ... [[-3]
-  ...  [-7]
-  ...  [-6]
-  ...  [-6]
-  ...  [-5]], shape=(5, 1), dtype=int16)
-  ```
+    ```python
+    >>> import tensorflow_io as tfio
+    >>>
+    >>> samples = tfio.IOTensor.from_audio("sample.wav")
+    >>> print(samples[1000:1005])
+    ... tf.Tensor(
+    ... [[-3]
+    ...  [-7]
+    ...  [-6]
+    ...  [-6]
+    ...  [-5]], shape=(5, 1), dtype=int16)
+    ```
 
-  ### Indexable vs. Iterable
+    ### Indexable vs. Iterable
 
-  While many IO formats are natually considered as iterable only, in most
-  of the situations they could still be accessed by indexing through certain
-  workaround. For example, a Kafka stream is not directly indexable yet the
-  stream could be saved in memory or disk to allow indexing. Another example
-  is the packet capture (PCAP) file in networking area. The packets inside
-  a PCAP file is concatenated sequentially. Since each packets could have
-  a variable length, the only way to access each packet is to read one
-  packet at a time. If the PCAP file is huge (e.g., hundreds of GBs or even
-  TBs), it may not be realistic (or necessarily) to save the index of every
-  packet in memory. We could consider PCAP format as iterable only.
+    While many IO formats are natually considered as iterable only, in most
+    of the situations they could still be accessed by indexing through certain
+    workaround. For example, a Kafka stream is not directly indexable yet the
+    stream could be saved in memory or disk to allow indexing. Another example
+    is the packet capture (PCAP) file in networking area. The packets inside
+    a PCAP file is concatenated sequentially. Since each packets could have
+    a variable length, the only way to access each packet is to read one
+    packet at a time. If the PCAP file is huge (e.g., hundreds of GBs or even
+    TBs), it may not be realistic (or necessarily) to save the index of every
+    packet in memory. We could consider PCAP format as iterable only.
 
-  As we could see the, availability of memory size could be a factor to decide
-  if a format is indexable or not. However, this factor could also be blurred
-  as well in distributed computing. One common case is the file format that
-  might be splittable where a file could be split into multiple chunks
-  (without read the whole file) with no data overlapping in between those
-  chunks. For example, a text file could be reliably split into multiple
-  chunks with line feed (LF) as the boundary. Processing of chunks could then
-  be distributed across a group of compute node to speed up (by reading small
-  chunks into memory). From that standpoint, we could still consider splittable
-  formats as indexable.
+    As we could see the, availability of memory size could be a factor to decide
+    if a format is indexable or not. However, this factor could also be blurred
+    as well in distributed computing. One common case is the file format that
+    might be splittable where a file could be split into multiple chunks
+    (without read the whole file) with no data overlapping in between those
+    chunks. For example, a text file could be reliably split into multiple
+    chunks with line feed (LF) as the boundary. Processing of chunks could then
+    be distributed across a group of compute node to speed up (by reading small
+    chunks into memory). From that standpoint, we could still consider splittable
+    formats as indexable.
 
-  For that reason our focus is `IOTensor` with convinience indexing and slicing
-  through `__getitem__()` method.
+    For that reason our focus is `IOTensor` with convinience indexing and slicing
+    through `__getitem__()` method.
 
-  ### Lazy Read
+    ### Lazy Read
 
-  One useful feature of `IOTensor` is the lazy read. Data inside a file is not
-  read into memory until needed. This could be convenient where only a small
-  segment of the data is needed. For example, a WAV file could be as big as
-  GBs but in many cases only several seconds of samples are used for training
-  or inference purposes.
+    One useful feature of `IOTensor` is the lazy read. Data inside a file is not
+    read into memory until needed. This could be convenient where only a small
+    segment of the data is needed. For example, a WAV file could be as big as
+    GBs but in many cases only several seconds of samples are used for training
+    or inference purposes.
 
-  While CPU memory is cheap nowadays, GPU memory is still considered as an
-  expensive resource. It is also imperative to fit data in GPU memory for
-  speed up purposes. From that perspective lazy read could be very helpful.
+    While CPU memory is cheap nowadays, GPU memory is still considered as an
+    expensive resource. It is also imperative to fit data in GPU memory for
+    speed up purposes. From that perspective lazy read could be very helpful.
 
-  ### Association of Meta Data
+    ### Association of Meta Data
 
-  While a file format could consist of mostly numeric data, in may situations
-  the meta data is important as well. For example, in audio file format the
-  sample rate is a number that is necessary for almost everything. Association
-  of the sample rate with the sample of int16 Tensor is more helpful,
-  especially in eager mode.
+    While a file format could consist of mostly numeric data, in may situations
+    the meta data is important as well. For example, in audio file format the
+    sample rate is a number that is necessary for almost everything. Association
+    of the sample rate with the sample of int16 Tensor is more helpful,
+    especially in eager mode.
 
-  Example:
+    Example:
 
-  ```python
-  >>> import tensorflow_io as tfio
-  >>>
-  >>> samples = tfio.IOTensor.from_audio("sample.wav")
-  >>> print(samples.rate)
-  ... 44100
-  ```
+    ```python
+    >>> import tensorflow_io as tfio
+    >>>
+    >>> samples = tfio.IOTensor.from_audio("sample.wav")
+    >>> print(samples.rate)
+    ... 44100
+    ```
 
-  ### Nested Element Structure
+    ### Nested Element Structure
 
-  The concept of `IOTensor` is not limited to a Tensor of single data type.
-  It supports nested element structure which could consists of many
-  components and complex structures. The exposed API such as `shape()` or
-  `dtype()` will display the shape and data type of an individual Tensor,
-  or a nested structure of shape and data types for components of a
-  composite Tensor.
+    The concept of `IOTensor` is not limited to a Tensor of single data type.
+    It supports nested element structure which could consists of many
+    components and complex structures. The exposed API such as `shape()` or
+    `dtype()` will display the shape and data type of an individual Tensor,
+    or a nested structure of shape and data types for components of a
+    composite Tensor.
 
-  Example:
+    Example:
 
-  ```python
-  >>> import tensorflow_io as tfio
-  >>>
-  >>> samples = tfio.IOTensor.from_audio("sample.wav")
-  >>> print(samples.shape)
-  ... (22050, 2)
-  >>> print(samples.dtype)
-  ... <dtype: 'int32'>
-  >>>
-  >>> features = tfio.IOTensor.from_json("feature.json")
-  >>> print(features.shape)
-  ... (TensorShape([Dimension(2)]), TensorShape([Dimension(2)]))
-  >>> print(features.dtype)
-  ... (tf.float64, tf.int64)
-  ```
+    ```python
+    >>> import tensorflow_io as tfio
+    >>>
+    >>> samples = tfio.IOTensor.from_audio("sample.wav")
+    >>> print(samples.shape)
+    ... (22050, 2)
+    >>> print(samples.dtype)
+    ... <dtype: 'int32'>
+    >>>
+    >>> features = tfio.IOTensor.from_json("feature.json")
+    >>> print(features.shape)
+    ... (TensorShape([Dimension(2)]), TensorShape([Dimension(2)]))
+    >>> print(features.dtype)
+    ... (tf.float64, tf.int64)
+    ```
 
-  ### Access Columns of Tabular Data Formats
+    ### Access Columns of Tabular Data Formats
 
-  May file formats such as Parquet or Json are considered as Tabular because
-  they consists of columns in a table. With `IOTensor` it is possible to
-  access individual columns through `__call__()`.
+    May file formats such as Parquet or Json are considered as Tabular because
+    they consists of columns in a table. With `IOTensor` it is possible to
+    access individual columns through `__call__()`.
 
-  Example:
+    Example:
 
-  ```python
-  >>> import tensorflow_io as tfio
-  >>>
-  >>> features = tfio.IOTensor.from_json("feature.json")
-  >>> print(features.shape("floatfeature"))
-  ... (2,)
-  >>> print(features.dtype("floatfeature"))
-  ... <dtype: 'float64'>
-  >>>
-  >>> print(features("floatfeature").shape)
-  ... (2,)
-  >>> print(features("floatfeature").dtype)
-  ... <dtype: 'float64'>
-  ```
+    ```python
+    >>> import tensorflow_io as tfio
+    >>>
+    >>> features = tfio.IOTensor.from_json("feature.json")
+    >>> print(features.shape("floatfeature"))
+    ... (2,)
+    >>> print(features.dtype("floatfeature"))
+    ... <dtype: 'float64'>
+    >>>
+    >>> print(features("floatfeature").shape)
+    ... (2,)
+    >>> print(features("floatfeature").dtype)
+    ... <dtype: 'float64'>
+    ```
 
-  ### Conversion from and to Tensor and Dataset
+    ### Conversion from and to Tensor and Dataset
 
-  When needed, `IOTensor` could be converted into a `Tensor` (through
-  `to_tensor()`, or a `tf.data.Dataset` (through `to_dataset()`, to
-  suppor operations that is only available through `Tensor` or
-  `tf.data.Dataset`.
+    When needed, `IOTensor` could be converted into a `Tensor` (through
+    `to_tensor()`, or a `tf.data.Dataset` (through `to_dataset()`, to
+    suppor operations that is only available through `Tensor` or
+    `tf.data.Dataset`.
 
-  Example:
+    Example:
 
-  ```python
-  >>> import tensorflow as tf
-  >>> import tensorflow_io as tfio
-  >>>
-  >>> features = tfio.IOTensor.from_json("feature.json")
-  >>>
-  >>> features_tensor = features.to_tensor()
-  >>> print(features_tensor())
-  ... (<tf.Tensor: id=21, shape=(2,), dtype=float64, numpy=array([1.1, 2.1])>,
-  ...  <tf.Tensor: id=22, shape=(2,), dtype=int64, numpy=array([2, 3])>)
-  >>>
-  >>> features_dataset = features.to_dataset()
-  >>> print(features_dataset)
-  ... <_IOTensorDataset shapes: ((), ()), types: (tf.float64, tf.int64)>
-  >>>
-  >>> dataset = tf.data.Dataset.zip((features_dataset, labels_dataset))
-  ```
+    ```python
+    >>> import tensorflow as tf
+    >>> import tensorflow_io as tfio
+    >>>
+    >>> features = tfio.IOTensor.from_json("feature.json")
+    >>>
+    >>> features_tensor = features.to_tensor()
+    >>> print(features_tensor())
+    ... (<tf.Tensor: id=21, shape=(2,), dtype=float64, numpy=array([1.1, 2.1])>,
+    ...  <tf.Tensor: id=22, shape=(2,), dtype=int64, numpy=array([2, 3])>)
+    >>>
+    >>> features_dataset = features.to_dataset()
+    >>> print(features_dataset)
+    ... <_IOTensorDataset shapes: ((), ()), types: (tf.float64, tf.int64)>
+    >>>
+    >>> dataset = tf.data.Dataset.zip((features_dataset, labels_dataset))
+    ```
 
-  """
+    """
 
     # =============================================================================
     # Graph mode
@@ -202,12 +201,12 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def graph(cls, dtype):
         """Obtain a GraphIOTensor to be used in graph mode.
 
-    Args:
-      dtype: Data type of the GraphIOTensor.
+        Args:
+          dtype: Data type of the GraphIOTensor.
 
-    Returns:
-      A class of `GraphIOTensor`.
-    """
+        Returns:
+          A class of `GraphIOTensor`.
+        """
         v = GraphIOTensor
         v._dtype = dtype  # pylint: disable=protected-access
         return v
@@ -220,20 +219,20 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_tensor(cls, tensor, **kwargs):
         """Converts a `tf.Tensor` into a `IOTensor`.
 
-    Examples:
+        Examples:
 
-    ```python
-    ```
+        ```python
+        ```
 
-    Args:
-      tensor: The `Tensor` to convert.
+        Args:
+          tensor: The `Tensor` to convert.
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    Raises:
-      ValueError: If tensor is not a `Tensor`.
-    """
+        Raises:
+          ValueError: If tensor is not a `Tensor`.
+        """
         with tf.name_scope(kwargs.get("name", "IOFromTensor")):
             return io_tensor_ops.TensorIOTensor(tensor, internal=True)
 
@@ -241,33 +240,35 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_audio(cls, filename, **kwargs):
         """Creates an `IOTensor` from an audio file.
 
-    The following audio file formats are supported:
-    - WAV
-    - OGG
+        The following audio file formats are supported:
+        - WAV
+        - Flac
+        - Vorbis
+        - MP3
 
-    Args:
-      filename: A string, the filename of an audio file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of an audio file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromAudio")):
-            return audio_io_tensor_ops.AudioIOTensor(filename, internal=True)
+            return audio_ops.AudioIOTensor(filename)
 
     @classmethod
     def from_json(cls, filename, **kwargs):
         """Creates an `IOTensor` from an json file.
 
-    Args:
-      filename: A string, the filename of an json file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of an json file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromJSON")):
             return json_io_tensor_ops.JSONIOTensor(
                 filename, mode=kwargs.get("mode", None), internal=True
@@ -277,27 +278,27 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_kafka(cls, topic, partition=0, servers=None, configuration=None, **kwargs):
         """Creates an `IOTensor` from a Kafka stream.
 
-    Args:
-      topic: A `tf.string` tensor containing topic subscription.
-      partition: A `tf.int64` tensor containing the partition, by default 0.
-      servers: An optional list of bootstrap servers, by default
-         `localhost:9092`.
-      configuration: An optional `tf.string` tensor containing
-        configurations in [Key=Value] format. There are three
-        types of configurations:
-        Global configuration: please refer to 'Global configuration properties'
-          in librdkafka doc. Examples include
-          ["enable.auto.commit=false", "heartbeat.interval.ms=2000"]
-        Topic configuration: please refer to 'Topic configuration properties'
-          in librdkafka doc. Note all topic configurations should be
-          prefixed with `configuration.topic.`. Examples include
-          ["conf.topic.auto.offset.reset=earliest"]
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          topic: A `tf.string` tensor containing topic subscription.
+          partition: A `tf.int64` tensor containing the partition, by default 0.
+          servers: An optional list of bootstrap servers, by default
+             `localhost:9092`.
+          configuration: An optional `tf.string` tensor containing
+            configurations in [Key=Value] format. There are three
+            types of configurations:
+            Global configuration: please refer to 'Global configuration properties'
+              in librdkafka doc. Examples include
+              ["enable.auto.commit=false", "heartbeat.interval.ms=2000"]
+            Topic configuration: please refer to 'Topic configuration properties'
+              in librdkafka doc. Note all topic configurations should be
+              prefixed with `configuration.topic.`. Examples include
+              ["conf.topic.auto.offset.reset=earliest"]
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromKafka")):
             return kafka_io_tensor_ops.KafkaIOTensor(
                 topic=topic,
@@ -311,14 +312,14 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_feather(cls, filename, **kwargs):
         """Creates an `IOTensor` from an feather file.
 
-    Args:
-      filename: A string, the filename of an feather file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of an feather file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromFeather")):
             return feather_io_tensor_ops.FeatherIOTensor(filename, internal=True)
 
@@ -326,18 +327,18 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_arrow(cls, table, spec=None, **kwargs):
         """Creates an `IOTensor` from a pyarrow.Table.
 
-    Args:
-      table: An instance of a `pyarrow.Table`.
-      spec: A dict of `dataset:tf.TensorSpec` or `dataset:dtype`
-        pairs that specify the dataset selected and the tf.TensorSpec
-        or dtype of the dataset. In eager mode the spec is probed
-        automatically. In graph mode spec has to be specified.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          table: An instance of a `pyarrow.Table`.
+          spec: A dict of `dataset:tf.TensorSpec` or `dataset:dtype`
+            pairs that specify the dataset selected and the tf.TensorSpec
+            or dtype of the dataset. In eager mode the spec is probed
+            automatically. In graph mode spec has to be specified.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromArrow")):
             return arrow_io_tensor_ops.ArrowIOTensor(table, spec=spec, internal=True)
 
@@ -345,14 +346,14 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_lmdb(cls, filename, **kwargs):
         """Creates an `IOTensor` from a LMDB key/value store.
 
-    Args:
-      filename: A string, the filename of a LMDB file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of a LMDB file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromLMDB")):
             return lmdb_io_tensor_ops.LMDBIOTensor(filename, internal=True)
 
@@ -360,18 +361,18 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_hdf5(cls, filename, spec=None, **kwargs):
         """Creates an `IOTensor` from an hdf5 file.
 
-    Args:
-      filename: A string, the filename of an hdf5 file.
-      spec: A dict of `dataset:tf.TensorSpec` or `dataset:dtype`
-        pairs that specify the dataset selected and the tf.TensorSpec
-        or dtype of the dataset. In eager mode the spec is probed
-        automatically. In graph mode spec has to be specified.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of an hdf5 file.
+          spec: A dict of `dataset:tf.TensorSpec` or `dataset:dtype`
+            pairs that specify the dataset selected and the tf.TensorSpec
+            or dtype of the dataset. In eager mode the spec is probed
+            automatically. In graph mode spec has to be specified.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromHDF5")):
             return hdf5_io_tensor_ops.HDF5IOTensor(filename, spec=spec, internal=True)
 
@@ -379,14 +380,14 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_csv(cls, filename, **kwargs):
         """Creates an `IOTensor` from an csv file.
 
-    Args:
-      filename: A string, the filename of an csv file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of an csv file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromCSV")):
             return csv_io_tensor_ops.CSVIOTensor(filename, internal=True)
 
@@ -394,15 +395,15 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_avro(cls, filename, schema, **kwargs):
         """Creates an `IOTensor` from an avro file.
 
-    Args:
-      filename: A string, the filename of an avro file.
-      schema: A string, the schema of an avro file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of an avro file.
+          schema: A string, the schema of an avro file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromAvro")):
             return avro_io_tensor_ops.AvroIOTensor(
                 filename, schema, internal=True, **kwargs
@@ -412,14 +413,14 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_ffmpeg(cls, filename, **kwargs):
         """Creates an `IOTensor` from a audio/video file.
 
-    Args:
-      filename: A string, the filename of a audio/video file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of a audio/video file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromFFmpeg")):
             return ffmpeg_io_tensor_ops.FFmpegIOTensor(filename, internal=True)
 
@@ -427,14 +428,14 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_parquet(cls, filename, **kwargs):
         """Creates an `IOTensor` from a parquet file.
 
-    Args:
-      filename: A string, the filename of a parquet file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of a parquet file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromParquet")):
             return parquet_io_tensor_ops.ParquetIOTensor(filename, internal=True)
 
@@ -442,16 +443,16 @@ class IOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-access
     def from_tiff(cls, filename, **kwargs):
         """Creates an `IOTensor` from a tiff file.
 
-    Note tiff file may consists of multiple images with different shapes.
+        Note tiff file may consists of multiple images with different shapes.
 
-    Args:
-      filename: A string, the filename of a tiff file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of a tiff file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromTIFF")):
             return tiff_io_tensor_ops.TIFFIOTensor(filename, internal=True)
 
@@ -467,24 +468,20 @@ class GraphIOTensor(io_tensor_ops._IOTensor):  # pylint: disable=protected-acces
     def from_audio(cls, filename, **kwargs):
         """Creates an `IOTensor` from an audio file.
 
-    The following audio file formats are supported:
-    - WAV
-    - OGG
+        The following audio file formats are supported:
+        - WAV
+        - Flac
+        - Vorbis
+        - MP3
 
-    Args:
-      filename: A string, the filename of an audio file.
-      dtype: Data type of the audio file.
-      name: A name prefix for the IOTensor (optional).
+        Args:
+          filename: A string, the filename of an audio file.
+          dtype: Data type of the audio file.
+          name: A name prefix for the IOTensor (optional).
 
-    Returns:
-      A `IOTensor`.
+        Returns:
+          A `IOTensor`.
 
-    """
+        """
         with tf.name_scope(kwargs.get("name", "IOFromAudio")):
-            resource = core_ops.io_audio_readable_init(filename)
-            dtype = cls._dtype
-            shape, _, rate = core_ops.io_audio_readable_spec(resource)
-            shape = tf.unstack(shape)
-            return audio_io_tensor_ops.AudioGraphIOTensor(
-                resource, shape, dtype, rate, internal=True
-            )
+            return audio_ops.AudioIOTensor(filename, dtype=cls._dtype)
